fix: check if comment exist on delete comment

+++ project/apps/blog/src/app/comments/comments.service.ts
@@ -1,26 +1,33 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateCommentDto } from './dto';
 import { CommentsRepository } from './repositories';
 import { CommentEntity } from './entities';
+import { CommentMessage } from './consts';
 
 @Injectable()
 export class CommentsService {
   constructor(private readonly commentsRepository: CommentsRepository) {}
 
-  createComment(dto: CreateCommentDto) {
+  async createComment(dto: CreateCommentDto) {
     const commentEntity = new CommentEntity({ ...dto, authorId: '123' });
     return this.commentsRepository.create(commentEntity);
   }
 
-  getAllComentsByPublication(publicationId: number) {
+  async getAllComentsByPublication(publicationId: number) {
     return this.commentsRepository.findByPublication(publicationId);
   }
 
-  getCommentById(commentId: number) {
+  async getCommentById(commentId: number) {
     return this.commentsRepository.findById(commentId);
   }
 
-  deleteComment(commentId: number) {
+  async deleteComment(commentId: number) {
+    const comment = await this.getCommentById(commentId);
+
+    if (!comment) {
+      throw new NotFoundException(CommentMessage.NotFound);
+    }
+
     return this.commentsRepository.destroy(commentId);
   }
 }
